Bot reading updates through subscribing to topic
