+++ src/main/scala/com/canoe/telegram/api/Bot.scala
@@ -2,15 +2,13 @@ package com.canoe.telegram.api
 
 import cats.Functor
 import cats.implicits._
-import com.canoe.telegram.clients.RequestHandler
-import com.canoe.telegram.methods.updates.GetUpdates
 import com.canoe.telegram.models.{InlineQuery, Message, Update}
 import fs2.Stream
 
-class Bot[F[_] : Functor](implicit client: RequestHandler[F]) {
+class Bot[F[_] : Functor](updateTopic: UpdatePolling[F]) {
   def messages: Stream[F, Message] =
     updates.map(_.message).collect {
-      case Some(message) => message
+      case Some(v) => v
     }
 
   def inlineQueries: Stream[F, InlineQuery] =
@@ -18,25 +16,7 @@ class Bot[F[_] : Functor](implicit client: RequestHandler[F]) {
       case Some(v) => v
     }
 
-  def updates: Stream[F, Update] = pollUpdates(0)
-
-  private def pollUpdates(startOffset: Long): Stream[F, Update] =
-    Stream(()).repeat.covary[F]
-      .evalMapAccumulate(startOffset) { case (offset, _) => requestUpdates(offset) }
-      .flatMap { case (_, updates) => Stream.emits(updates) }
-
-  private def requestUpdates(offset: Long): F[(Long, List[Update])] =
-    client
-      .execute(GetUpdates(Some(offset)))
-      .map(_.toList)
-      .map(updates => (lastOffset(updates).map(_ + 1).getOrElse(offset),  updates))
-
-
-  private def lastOffset(updates: List[Update]): Option[Long] =
-    updates match {
-      case Nil => None
-      case nonEmpty => Some(nonEmpty.map(_.updateId).max)
-    }
+  def updates: Stream[F, Update] = updateTopic.fork
 }
