+++ src/main/scala/com/canoe/telegram/clients/SttpClient.scala
@@ -27,6 +27,7 @@ class SttpClient[F[_]](token: String, telegramHost: String = "api.telegram.org")
 
   override def sendRequest[R, T <: Request[_]](request: T)(implicit encT: Encoder[T], decR: Decoder[R]): F[R] = {
     val url = apiBaseUrl + request.methodName
+    println(s"Calling $url")
 
     val sttpRequest: RequestT[Id, String, Nothing] = request match {
       case r: JsonRequest[_] =>
